+++ src/webpimage.cpp
@@ -348,7 +348,7 @@ namespace Exiv2 {
                     if (outIo.write((const byte*)WEBP_CHUNK_HEADER_ICCP, WEBP_TAG_SIZE) != WEBP_TAG_SIZE) throw Error(21);
                     ul2Data(data, (uint32_t) iccProfile_.size_, littleEndian);
                     if (outIo.write(data, WEBP_TAG_SIZE) != WEBP_TAG_SIZE) throw Error(21);
-                    if (outIo.write(iccProfile_.pData_, (long)iccProfile_.size_) != (long)iccProfile_.size_) {
+                    if (outIo.write(iccProfile_.pData_, iccProfile_.size_) != iccProfile_.size_) {
                         throw Error(21);
                     }
                     has_icc = false;
@@ -597,21 +597,21 @@ namespace Exiv2 {
                 bool  be_header = false;
                 long  pos = -1;
 
-                pos = getHeaderOffset (payload.pData_, (long)payload.size_, (byte*)&exifLongHeader, 4);
+                pos = getHeaderOffset (payload.pData_, payload.size_, (byte*)&exifLongHeader, 4);
                 if (pos == -1) {
-                    pos = getHeaderOffset (payload.pData_, (long)payload.size_, (byte*)&exifLongHeader, 6);
+                    pos = getHeaderOffset (payload.pData_, payload.size_, (byte*)&exifLongHeader, 6);
                     if (pos != -1) {
                         s_header = true;
                     }
                 }
                 if (pos == -1) {
-                    pos = getHeaderOffset (payload.pData_, (long)payload.size_, (byte*)&exifTiffLEHeader, 3);
+                    pos = getHeaderOffset (payload.pData_, payload.size_, (byte*)&exifTiffLEHeader, 3);
                     if (pos != -1) {
                         le_header = true;
                     }
                 }
                 if (pos == -1) {
-                    pos = getHeaderOffset (payload.pData_, (long)payload.size_, (byte*)&exifTiffBEHeader, 4);
+                    pos = getHeaderOffset (payload.pData_, payload.size_, (byte*)&exifTiffBEHeader, 4);
                     if (pos != -1) {
                         be_header = true;
                     }
@@ -640,7 +640,7 @@ namespace Exiv2 {
                     memcpy(rawExifData + 6, (char*)&exifShortHeader, 6);
                 }
 
-                memcpy(rawExifData + offset, payload.pData_, (long)payload.size_);
+                memcpy(rawExifData + offset, payload.pData_, payload.size_);
 
 #ifdef DEBUG
                 std::cout << "Display Hex Dump [size:" << (unsigned long)size << "]" << std::endl;
